--- C Programa Principal/Bd_Sensores/sensores_actuador.sql
@@ -1,55 +1,0 @@
-CREATE DATABASE  IF NOT EXISTS `sensores` /*!40100 DEFAULT CHARACTER SET utf8mb4 COLLATE utf8mb4_0900_ai_ci */ /*!80016 DEFAULT ENCRYPTION='N' */;
-USE `sensores`;
--- MySQL dump 10.13  Distrib 8.0.36, for Win64 (x86_64)
---
--- Host: 127.0.0.1    Database: sensores
--- ------------------------------------------------------
--- Server version	8.0.37
-
-/*!40101 SET @OLD_CHARACTER_SET_CLIENT=@@CHARACTER_SET_CLIENT */;
-/*!40101 SET @OLD_CHARACTER_SET_RESULTS=@@CHARACTER_SET_RESULTS */;
-/*!40101 SET @OLD_COLLATION_CONNECTION=@@COLLATION_CONNECTION */;
-/*!50503 SET NAMES utf8 */;
-/*!40103 SET @OLD_TIME_ZONE=@@TIME_ZONE */;
-/*!40103 SET TIME_ZONE='+00:00' */;
-/*!40014 SET @OLD_UNIQUE_CHECKS=@@UNIQUE_CHECKS, UNIQUE_CHECKS=0 */;
-/*!40014 SET @OLD_FOREIGN_KEY_CHECKS=@@FOREIGN_KEY_CHECKS, FOREIGN_KEY_CHECKS=0 */;
-/*!40101 SET @OLD_SQL_MODE=@@SQL_MODE, SQL_MODE='NO_AUTO_VALUE_ON_ZERO' */;
-/*!40111 SET @OLD_SQL_NOTES=@@SQL_NOTES, SQL_NOTES=0 */;
-
---
--- Table structure for table `actuador`
---
-
-DROP TABLE IF EXISTS `actuador`;
-/*!40101 SET @saved_cs_client     = @@character_set_client */;
-/*!50503 SET character_set_client = utf8mb4 */;
-CREATE TABLE `actuador` (
-  `id_actuador` int NOT NULL AUTO_INCREMENT,
-  `estado` varchar(10) NOT NULL,
-  `fecha_accion` datetime DEFAULT CURRENT_TIMESTAMP,
-  `alerta` varchar(100) DEFAULT NULL,
-  PRIMARY KEY (`id_actuador`)
-) ENGINE=InnoDB AUTO_INCREMENT=8 DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;
-/*!40101 SET character_set_client = @saved_cs_client */;
-
---
--- Dumping data for table `actuador`
---
-
-LOCK TABLES `actuador` WRITE;
-/*!40000 ALTER TABLE `actuador` DISABLE KEYS */;
-INSERT INTO `actuador` VALUES (1,'ON','2024-10-05 15:27:48','Alerta: Temperatura fuera de rango - 32.0 °C'),(2,'ON','2024-10-05 15:27:49','Alerta: Humedad fuera de rango - 28.0%'),(3,'OFF','2024-10-05 15:57:23','Temperatura dentro de rango'),(4,'OFF','2024-10-05 15:57:23','Temperatura dentro de rango'),(5,'OFF','2024-10-05 15:57:23','Temperatura dentro de rango'),(6,'OFF','2024-10-05 15:57:23','Temperatura dentro de rango'),(7,'ON','2024-10-05 15:57:23','Alerta: Temperatura fuera de rango - 32.0 °C');
-/*!40000 ALTER TABLE `actuador` ENABLE KEYS */;
-UNLOCK TABLES;
-/*!40103 SET TIME_ZONE=@OLD_TIME_ZONE */;
-
-/*!40101 SET SQL_MODE=@OLD_SQL_MODE */;
-/*!40014 SET FOREIGN_KEY_CHECKS=@OLD_FOREIGN_KEY_CHECKS */;
-/*!40014 SET UNIQUE_CHECKS=@OLD_UNIQUE_CHECKS */;
-/*!40101 SET CHARACTER_SET_CLIENT=@OLD_CHARACTER_SET_CLIENT */;
-/*!40101 SET CHARACTER_SET_RESULTS=@OLD_CHARACTER_SET_RESULTS */;
-/*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
-/*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
-
--- Dump completed on 2024-10-28 20:03:07
