+++ C Programa Principal/Visualizador-Temp-hum/Codigo Python/insertar_datos.py
