+++ C Programa Principal/Visualizador-Temp-hum/Programa-ESP32/.gitignore
