--- C Programa Principal/Visualizador_Temp_Hum/platformio.ini
@@ -1,21 +1,0 @@
-; PlatformIO Project Configuration File
-;
-;   Build options: build flags, source filter
-;   Upload options: custom upload port, speed and extra flags
-;   Library options: dependencies, extra library storages
-;   Advanced options: extra scripting
-;
-; Please visit documentation for the other options and examples
-; https://docs.platformio.org/page/projectconf.html
-
-[env:esp32dev]
-platform = espressif32
-board = esp32dev
-framework = arduino
-monitor_speed = 115200
-lib_deps = 
-    adafruit/DHT sensor library @ ^1.4.4
-    adafruit/Adafruit Unified Sensor @ ^1.1.9
-    adafruit/Adafruit NeoPixel @ ^1.10.6
-	https://github.com/ChuckBell/MySQL_Connector_Arduino
-
