--- C Programa Principal/test Temperatura y Humedad/Programa-ESP32/.gitignore
@@ -1,5 +1,0 @@
-.pio
-.vscode/.browse.c_cpp.db*
-.vscode/c_cpp_properties.json
-.vscode/launch.json
-.vscode/ipch
